Gallows ImageView stuff. Currently not working but no errors
I added all the stuff to playscreen.  There are a few problems.  I tried
adjusting the text size of the text in the TextView that displays the
word but if I do then the keyboard doesn't show up at all.  I don't know
how to deal with that.  I will figure out the ImageView problems I'm
having.  All we need to do is run switchImage when the wrong letter is
clicked and it will work (when I fix it)